+++ src/main/java/gregtech/common/items/behaviors/CoverDigitalInterfaceWirelessPlaceBehaviour.java
@@ -2,7 +2,7 @@ package gregtech.common.items.behaviors;
 
 import gregtech.api.cover.CoverDefinition;
 import gregtech.api.metatileentity.MetaTileEntityHolder;
-import gregtech.client.renderer.handler.WorldRenderEventRenderer;
+import gregtech.client.renderer.handler.BlockPosHighlightRenderer;
 import gregtech.common.metatileentities.multi.electric.centralmonitor.MetaTileEntityCentralMonitor;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.entity.Entity;
@@ -40,7 +40,7 @@ public class CoverDigitalInterfaceWirelessPlaceBehaviour extends CoverPlaceBehav
             if (held == itemStack) {
                 BlockPos pos = getRemotePos(itemStack);
                 if (pos != null) {
-                    WorldRenderEventRenderer.renderBlockBoxHighLight(pos, 1500);
+                    BlockPosHighlightRenderer.renderBlockBoxHighLight(pos, 1500);
                 }
             }
         }
